online export/import

+++ modules/system/src/main/sh/bin/entrypoint-wrapper.sh
@@ -17,9 +17,24 @@ modify_keycloak_start_cmd() {
 graceful_shutdown() {
 	echo  "SIGTERM received, shutting down Keycloak gracefully..."
 
-	# Send SIGTERM to Keycloak process
-	kill  -TERM "$keycloak_pid"
-	wait  "$keycloak_pid"
+	pid_file="$KC_PID_FILE"
+
+	if [ -f "$pid_file" ]; then
+	    keycloak_pid=$(cat "$pid_file")
+
+	    if ! [[ $keycloak_pid =~ ^[0-9]+$ ]]; then
+		echo "Invalid PID: $keycloak_pid"
+		exit 1
+	    fi
+
+	    kill -TERM "$keycloak_pid"
+
+	    wait "$keycloak_pid"
+	else
+	    echo "PID file not found: $pid_file"
+	    exit 1
+	fi
+
 
 	if  [ -f "$EXPORT_USERS_SCRIPT" ]; then
 		# KEYCLOAK_BACKUP_ON_SIGTERM is an custom environment variable designed for use in Dockerized Keycloak environments,
@@ -64,6 +79,5 @@ modify_keycloak_start_cmd
 echo "Starting Keycloak with command: $KEYCLOAK_START_CMD"
 $KEYCLOAK_START_CMD &
 keycloak_pid=$!
-
-# Wait for Keycloak process to finish
+echo $keycloak_pid > "$KC_PID_FILE"
 wait $keycloak_pid
